GET Route and Controller overhaul
Greatly modified the functionality of the GET routes for user and auction. It now has optional parameters, and will query the database based on those parameters. If no parameters are given, it will return the first user/auction it finds. Additionally, when querying based on displayName or auctionName it will use regex to find similar matches.